backend/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
dotmap==1.3.14
eventlet==0.25.2
Flask==1.1.2
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SocketIO==4.3.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
python-engineio==3.12.1
python-socketio==4.5.1
PyYAML==5.3.1
redis==3.5.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
Werkzeug==0.16.0
